+++ src/Symfony/Component/Validator/Tests/Constraints/LessThanValidatorWithNegativeConstraintTest.php
@@ -111,4 +111,9 @@ class LessThanValidatorWithNegativeConstraintTest extends LessThanValidatorTest
     {
         $this->markTestSkipped('PropertyPath option is not used in Positive constraint');
     }
+
+    public function testInvalidComparisonToPropertyPathAddsPathAsParameter()
+    {
+        $this->markTestSkipped('PropertyPath option is not used in Negative constraint');
+    }
 }