+++ src/Symfony/Bundle/TwigBundle/CHANGELOG.md
@@ -9,6 +9,12 @@ CHANGELOG
  * removed `TwigEngine` class, use `Twig\Environment` instead
  * removed `FilesystemLoader` and `NativeFilesystemLoader`, use Twig notation for templates instead
 
+4.4.0
+-----
+
+ * marked the `TemplateIterator` as `internal`
+ * added HTML comment to beginning and end of `exception_full.html.twig`
+
 4.2.0
 -----